protect deleteUser

+++ backend/controllers/users.js
@@ -99,29 +99,32 @@ const updateUser = asyncWrapper(async (req, res, next) => {
 })
 
 const deleteUser = asyncWrapper(async (req, res, next) => {
-  const user_auth0Id = req.params.id
+  const userTokenId = req.auth.payload.sub;
+  const auth0Id = req.params.id
 
-  // get user from database
-  const user = await User.findOne({ auth0Id: user_auth0Id })
-  if (!user) {
-    return next(createCustomError(`Error (find): No user with id: ${user_auth0Id}`, 404), req, res)
+  // get user from database (need user._id to delete workoutList)
+  const user = await User.findOne({ auth0Id })
+
+  if (!user || userTokenId != user.auth0Id) {
+    return next(createCustomError(`No user with auth0Id: ${auth0Id}`, 404), req, res)
   }
 
   // delete user from database
-  const deletedUser = await User.findOneAndDelete({ auth0Id: user_auth0Id })
+  const deletedUser = await User.findOneAndDelete({ auth0Id })
+
   if (!deletedUser) {
-    return next(createCustomError(`Error (delete): No user with id: ${user_auth0Id}`, 404), req, res)
+    return next(createCustomError(`No user with auth0Id: ${auth0Id}`, 404), req, res)
   }
 
   // delete user workouts from database 
   const workoutList = await Workout.findOneAndDelete({ userId: user._id })
   if (!workoutList) {
-    return next(createCustomError(`No workoutList for id: ${user_auth0Id}`, 404), req, res)
+    return next(createCustomError(`No workoutList for id: ${auth0Id}`, 404), req, res)
   }
 
-  deleteAuthUser(user_auth0Id)
+  deleteAuthUser(auth0Id)
 
-  res.status(200).json({ msg: 'deleted user' })
+  res.status(200).json({ msg: 'User Succesfully Deleted' })
 })
 
 const getSingleUser = asyncWrapper(async (req, res, next) => {
